Updated changelog.md and troubleshooting.md with new version and new known fail case and solution.
Moved additional features to its own section.

Fixed issue in meshing script where it would look for a temporary file when it didn't need to.

Added templates and guidelines.

Documentation and minor file fix.
Added version 1.0.1 and 1.0.2 to version changelog.
Fixed issue in CITATION.cff about repository code.

+++ mesh_libs/A_IA_mesher.py
@@ -454,7 +454,7 @@ def run_icem_apdl():
     except Exception as e:
         print(f'Error: {e}\n ANSYS ICEM has failed to run please check you have files in the correct location')
 
-    if temp_file:
+    if gui_ins.bm_rot and temp_file:
         try:
             os.remove(temp_file.name)
         except OSError:
